requirements.txt
telethon==1.36.0
python-dotenv==1.0.1
schedule==1.2.2
requests==2.31.0
